==> notebooks/01_loading.ipynb <==
### Chargement du dataset

##### 1. Importer le dataset `7Xan7der7/us_airline_sentiment`

from datasets import load_dataset

dataset = load_dataset('7Xan7der7/us_airline_sentiment')

print(dataset)

##### 2. Convertir les données en Pandas dataframes

import pandas as pd

data = pd.DataFrame(dataset['train'])

data.head()

##### 3. Sauvegarder les données

data.to_csv('../data/raw/data.csv', index=False)

print('Les données ont été enregistrées avec succès !')
