==> notebooks/03_preprocessing.ipynb <==
### Nettoyage et normalisation des données

#### 1. Charger les données

import pandas as pd

data = pd.read_csv('../data/raw/data.csv')

print('Les données ont été chargées')

#### 2. Supprimer les doublons

inital_len = len(data)

data = data.drop_duplicates()

current_lignes = len(data)

print(f"{inital_len - current_lignes} lignes ont été supprimées. {current_lignes} restants")

##### 3. Gérer les valeurs manquantes

- Identifier les valeurs manquantes

data.isna().sum()

- Garder les colonnes pertinents

data = data[['airline_sentiment', 'airline_sentiment_confidence', 'airline', 'text']]

##### 4. Traitement et Nettoyage de données

data["text"] = data["text"].fillna("").str.strip()                  # Debut et fin 
data["text"] = data["text"].str.replace(r"\s+", " ", regex=True)    # Espaces doublons

##### 4.1. Data 1: data_with_mentions

- Sauvegarder les données

data.to_csv('../data/processed/data_with_mentions.csv', index=False)

##### 4.2. Data 2: data_clean

- Supprimer les mentions 

data_2 = data.copy()

data_2["text"] = data_2["text"].str.replace(r'@[^\s]+', '', regex=True)

- Supprimer les espaces supplémentaires

data_2["text"] = data_2["text"].fillna("").str.strip()
data_2["text"] = data_2["text"].str.replace(r"\s+", " ", regex=True)

- Sauvegarder les données propres

data_2.to_csv('../data/processed/data_clean.csv', index=False)

##### 4.3. Data 3: data_balanced_undersampling

- Equilibrer les classes (UnderSampling)

data_3 = data.copy()

positive_class = data_3[data_3['airline_sentiment'] == 'positive']
neutral_class = data_3[data_3['airline_sentiment'] == 'neutral']
negative_class = data_3[data_3['airline_sentiment'] == 'negative']

print('--- Avant:')
print('\tPositive:', len(positive_class))
print('\tNeutral: ', len(neutral_class))
print('\tNegative:', len(negative_class))

frac = len(neutral_class) / len(negative_class)
balanced_negative = negative_class.sample(frac=frac)

print('\n--- Après:')
print('\tPositive:', len(positive_class))
print('\tNeutral: ', len(neutral_class))
print('\tNegative:', len(balanced_negative))

data_3 = positive_class.merge(neutral_class, 'outer').merge(balanced_negative, 'outer')

print('\n--- Total:', len(data_3))


- Sauvegarder les données propres

data_3.to_csv('../data/processed/data_balanced_undersampling.csv', index=False)
