==> notebooks/04_embeddings.ipynb <==
### Génération des embeddings

##### 1. Charger les données

import pandas as pd

data_clean = pd.read_csv('../data/processed/data_clean.csv')
data_with_mentions = pd.read_csv('../data/processed/data_with_mentions.csv')
data_balanced_undersampling = pd.read_csv('../data/processed/data_balanced_undersampling.csv')

##### 2. Séparer les données en ensembles d'entraînement et de test

from sklearn.model_selection import train_test_split

def divide_sets(data):
    X = data.drop('airline_sentiment', axis=1)
    y = data['airline_sentiment']

    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.15, random_state=42, stratify=y)

    print('X_Train:', len(X_train))
    print('X_Test :', len(X_test))
    print('Y_Train:', len(y_train))
    print('Y_Test :', len(y_test))

    return [X_train, X_test, y_train, y_test]

- Data 1: data_clean

X_train_clean, X_test_clean, y_train_clean, y_test_clean = \
divide_sets(data_clean)

- Data 2: data_with_mentions

X_train_mentions, X_test_mentions, y_train_mentions, y_test_mentions = \
divide_sets(data_with_mentions)

- Data 3: data_balanced_undersampling

X_train_undersampling, X_test_undersampling, y_train_undersampling, y_test_undersampling = \
divide_sets(data_balanced_undersampling)

##### 3. Le choix de modèles de génération d’embeddings

Parmi les meilleurs modèles de génération d’embeddings, on trouve :

| Modèle              | Qualité         | Multilingue       | Vitesse |
| ------------------- | --------------- | ----------------- | ------- |
| BGE-large-m3        | Excellent       | Oui               | Lent |
| E5-large-v2         | Excellent       | EN only           | Moyen |
| Multilingual-E5     | Bien            | Oui               | Moyen |
| GTE-large           | Bien            | EN only           | Rapide |
| MiniLM-L12-v2       | qualité faible  | Oui               | Très rapide |


On va utiliser `E5-large-v2` pour trouver un équilibre entre la performance et la vitesse d’entraînement.

##### 4. Prétraitement des textes pour le modéle `E5-large-v2`

Le modèle `E5-large-v2` a été entraîné sur du texte naturel (phrases réelles, majuscules, ponctuation, etc.), donc il faut garder le texte aussi proche que possible de sa forme originale,
tout en nettoyant les erreurs mineures ou les incohérences.

il est recommander d'éviter de:
- Supprimer les stopwords: Le modèle comprend leur rôle sémantique.
- Lemmatisation ou stemming: Cela modifie la structure linguistique et le sens.
- Supprimer la ponctuation: La ponctuation aide à comprendre le contexte.
- Tout mettre en minuscules (Sauf si le texte contient des majuscules aléatoires ou tout en majuscules.): Peut faire perdre des distinctions importantes (Apple ≠ apple, US ≠ us).
- Supprimer les chiffres: Les nombres ont souvent une signification utile (dates, montants…).

Il est Recommander (et parfois obligatoire) de:
- Supprimer les espaces inutiles: Nettoyer les débuts/fin de lignes et les espaces multiples.
- Gérer les valeurs manquantes: Remplacer les `NaN` ou textes vides par une chaîne vide.
- Ajouter le préfixe (Obligatoire): **"passage: "** pour les documents, **"query: "** pour les requêtes.
- Tronquer les textes très longs: Si un texte dépasse la limite (512 tokens environ), on peut le couper.

##### 5. Charger le modèle `intfloat/e5-large-v2` avec Sentence Transformers

from sentence_transformers import SentenceTransformer

model = SentenceTransformer('intfloat/e5-large-v2')

##### 6. Générer les embeddings pour les données

import numpy as np

def generate_embeddings(X_train, X_test, suffix):

    # Générer les embeddings d'entrainement

    texts = X_train['text'].tolist()

    train_embeddings = model.encode(
        texts,
        convert_to_numpy=True,
        normalize_embeddings=True,
        show_progress_bar=True
    )

    # Générer les embeddings de test

    texts = X_test['text'].tolist()

    test_embeddings = model.encode(
        texts,
        convert_to_numpy=True,
        normalize_embeddings=True,
        show_progress_bar=True
    )

    # Sauvegarder les embeddings Générés (.npy)

    np.save(f'../data/embedding/train_embeddings_{suffix}.npy', train_embeddings)
    np.save(f'../data/embedding/test_embeddings_{suffix}.npy', test_embeddings)

    return [train_embeddings, test_embeddings]

- Data 1: data_clean

train_embeddings_clean, test_embeddings_clean = \
generate_embeddings(X_train_clean, X_test_clean, 'clean')

- Data 2: data_with_mentions

train_embeddings_mentions, test_embeddings_mentions = \
generate_embeddings(X_train_mentions, X_test_mentions, 'mentions')

- Data 3: data_balanced_undersampling

train_embeddings_undersampling, test_embeddings_undersampling = \
generate_embeddings(X_train_undersampling, X_test_undersampling, 'undersampling')

##### 7. Afficher les embeddings Générés

print('Train - 1:', train_embeddings_clean.shape)
print('Test  - 1:', test_embeddings_clean.shape, '\n')
print('Train - 2:', train_embeddings_mentions.shape)
print('Test  - 2:', test_embeddings_mentions.shape, '\n')
print('Train - 3:', train_embeddings_undersampling.shape)
print('Test  - 3:', test_embeddings_undersampling.shape, '\n')

print(train_embeddings_clean)

##### 8. Sauvegarder les labels et identifiants

from sklearn.preprocessing import LabelEncoder
import joblib as jb

def encode_labels(y_train, y_test, suffix):

    encoder = LabelEncoder()

    # Metadata de l'entrainement

    train_metadata = pd.DataFrame()
    train_metadata['label_name'] = y_train
    train_metadata['id'] = range(0, len(train_metadata))
    train_metadata['label'] = encoder.fit_transform(train_metadata['label_name'])

    # Metadata de test

    test_metadata = pd.DataFrame()
    test_metadata['label_name'] = y_test
    test_metadata['id'] = range(0, len(test_metadata))
    test_metadata['label'] = encoder.transform(test_metadata['label_name'])

    # Sauvegarder les metadatas

    train_metadata.to_csv(f'../data/metadata/train_metadata_{suffix}.csv', index=False)
    test_metadata.to_csv(f'../data/metadata/test_metadata_{suffix}.csv', index=False)

    # Sauvegarder l'encoder

    jb.dump(encoder, f'../models/encoders/encoder_{suffix}.pkl')

    # Afficher un message de réussi

    print("Les métadonnées ont été générées et enregistrées.")


- Data 1: data_clean

encode_labels(y_train_clean, y_test_clean, 'clean')

- Data 2: data_with_mentions

encode_labels(y_train_mentions, y_test_mentions, 'mentions')

- Data 3: data_balanced_undersampling

encode_labels(y_train_undersampling, y_test_undersampling, 'undersampling')
